gemini-backend/requirements.txt
Flask
Flask-Cors
python-dotenv
google-generativeai
gunicorn
